+++ app/src/main/java/net/romanitalian/moneytrackerapp/TransactionsFragment.java
@@ -6,14 +6,47 @@ import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ListView;
+
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.List;
+import java.util.Locale;
 
 /**
  * Created by romadin on 27.04.2015.
  */
 public class TransactionsFragment extends Fragment {
+    private ListView listView;
+    private TransactionAdapter transactionAdapter;
+    List<Transactions> data = new ArrayList<>();
+    String date_format = "yyyy-MM-dd";
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return super.onCreateView(inflater, container, savedInstanceState);
+        View inflate = inflater.inflate(R.layout.fragment_transactions, container);
+        super.onCreateView(inflater, container, savedInstanceState);
+
+        List<Transactions> adapterData = getDataList();
+        transactionAdapter = new TransactionAdapter(getActivity(), adapterData);
+
+        listView = (ListView) getActivity().findViewById(R.id.list_view_id);
+        listView.setAdapter(transactionAdapter);
+        return inflate;
+    }
+
+    private List<Transactions> getDataList() {
+        DateFormat df = new SimpleDateFormat(date_format, new Locale("ru"));
+        Date now_calendar = Calendar.getInstance().getTime();
+        String now = df.format(now_calendar);
+        data.add(new Transactions("Huawei", "9800", now));
+        data.add(new Transactions("SamsungS3", "13000", now));
+        data.add(new Transactions("T-shirt", "300", now));
+        data.add(new Transactions("Jeans", "1500", now));
+        data.add(new Transactions("Printer", "4500", now));
+        return data;
     }
 }
