+++ app/src/main/res/layout/fragment_transactions.xml
@@ -1,0 +1,12 @@
+<RelativeLayout
+    style="@style/Container.Background"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <fragment
+        android:id="@+id/fragment_transactions_id"
+        class="net.romanitalian.moneytrackerapp.TransactionsFragment"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        />
+
+</RelativeLayout>
